Struct pgs_files::shadow::ShadowEntry

pub struct ShadowEntry {
    pub name: String,
    pub passwd: String,
    pub last_change: c_long,
    pub min: c_long,
    pub max: c_long,
    pub warning: c_long,
    pub inactivity: c_long,
    pub expires: c_long,
    pub flag: c_ulong,
}

An entry from /etc/shadow

Fields

name

Login name

passwd

Encrypted password

last_change

Date of last change (measured in days since 1970-01-01 00:00:00 +0000 (UTC))

min

Min number of days between changes

max

Max number of days between changes

warning

Number of days before password expires to warn user to change it

inactivity

Number of days after password expires until account is disabled

expires

Date when account expires (measured in days since 1970-01-01 00:00:00 +0000 (UTC))

flag

Reserved